apt/requirements_bootstrap.txt
python-debian

